remove tslint confic
remove ds_store

mv

stuff

more stuff

clean

+++ src/commands/generateProject.ts
@@ -1,46 +1,46 @@
 import * as vscode from "vscode";
+import * as path from "path";
 import * as prompts from "../util/vscodePrompts";
 import { INPUT_YAML_OPTIONS, GENERATOR_JAR_PATH } from "../constants";
 import * as fileUtil from "../util/file";
-import * as path from "path";
-import * as childProcess from "child_process";
-import { promisify } from "util";
-
-const exec = promisify(childProcess.exec);
+import * as processUtil from "../util/process";
 
 export async function generateProject(clickedFileUri: vscode.Uri | undefined): Promise<void> {
+  // extension uses a tmp directory to download / generate files
+  let tmpDirPath: string | undefined;
+
   try {
+    const inputYamlMethod = await prompts.askForYamlInputMethod();
+
     let yamlInputFileURI: vscode.Uri | undefined;
     let yamlInputURL: string | undefined;
 
-    const inputYamlMethod = await prompts.specifyYamlInputMethod();
-
     if (inputYamlMethod === INPUT_YAML_OPTIONS.FROM_FILE) {
-      yamlInputFileURI = await prompts.specifyYamlFile();
+      yamlInputFileURI = await prompts.askForYamlFile();
     } else if (inputYamlMethod === INPUT_YAML_OPTIONS.FROM_URL) {
-      yamlInputURL = await prompts.specifyYamlURL();
+      yamlInputURL = await prompts.askForYamlURL();
     }
 
-    // if neither an input FileURI or input URL are specified exit the generator
+    // if neither an input file URI or input URL are specified exit the generator
     if (yamlInputFileURI === undefined && yamlInputURL === undefined) {
       return;
     }
 
     // ask for a folder to generate rest client into
-    // use the folder clicked on by the user as the default if the command was
-    // triggered by the file explorer
-    const targetDirectory = await prompts.specifyTargetFolder(clickedFileUri);
+    // use the fileURI clicked on by the user as the default if
+    // the command was triggered from the file explorer and not the command palette
+    const targetDirectory = await prompts.askForTargetFolder(clickedFileUri);
     if (targetDirectory === undefined) {
       return;
     }
 
-    const packageName = await prompts.specifyPackageName(targetDirectory.fsPath);
+    const packageName = await prompts.askForPackageName(targetDirectory.fsPath);
     if (packageName === undefined) {
       return;
     }
 
     // make a tmp directory in the target folder to generate files into
-    const tmpDirPath = await fileUtil.generateTempDirectory(targetDirectory.fsPath);
+    tmpDirPath = await fileUtil.generateTempDirectory(targetDirectory.fsPath);
 
     // if they are using a URL download the file to the temp directory
     if (inputYamlMethod === INPUT_YAML_OPTIONS.FROM_URL) {
@@ -70,26 +70,26 @@ export async function generateProject(clickedFileUri: vscode.Uri | undefined): P
       yamlInputFileURI.fsPath +
       " -g java --library microprofile -o " +
       tmpDirPath +
-      "\
-          --api-package " +
+      " --api-package " +
       packageName +
-      ".api \
-          --model-package " +
+      ".api --model-package " +
       packageName +
       ".models";
 
-    await exec(jarCommand);
+    await processUtil.exec(jarCommand);
 
     const packagePath = packageName.replace(/\./g, path.sep);
     const generatedRestClientPath = path.resolve(tmpDirPath, "src", "main", "java", packagePath);
 
-    // copy the api/models folder into the target directory
+    // copy the api/models folder from the generated directory into the target directory
     await fileUtil.copy(generatedRestClientPath, targetDirectory.fsPath);
-
-    // remove the tmp directory after
-    await fileUtil.deleteDirectory(tmpDirPath);
   } catch (e) {
     console.error(e);
-    await vscode.window.showErrorMessage("Failed to generate rest client");
+    await vscode.window.showErrorMessage("Failed to generate MicroProfile rest client");
+  } finally {
+    // remove the tmp directory after if it exists
+    if (tmpDirPath !== undefined && (await fileUtil.exists(tmpDirPath))) {
+      await fileUtil.deleteDirectory(tmpDirPath);
+    }
   }
 }
